Add files via upload
This is an excel spreadsheet that details the logic behind if a servo needs to switch a track or not, to reach the desired location, and if so, does the track need to be "open" or "closed"
(Assuming closed is the default position)

This is also sorted depending on the current direction of the train. Which is important as knowing that information lets us cut the required array searching down. 

Signed-off-by: Eagleeyevortex <125773177+Eagleeyevortex@users.noreply.github.com>