V 1.07
In this version the logic for cases 0,1,2,3,7,14 are written (not yet tested) started work on case 4 and 8.  Move over, The 3 element in the sensor_states[15][2] array is now fixed and can be used to get the data for which track any given ir sensor is on. to get the sensor_state element working the function where values are fed to this array  in the beginning was updated with the new track values. Finally, one more update was made in this code and that is the trains can be placed anywhere on the track setup instead of the given positions in the perfect world scenario and the code will still work and update the arrays accordingly for the first few times.

Signed-off-by: Aryan Chavan <70076439+ChavAr24@users.noreply.github.com>